+++ ui/src/common/index.js
@@ -25,6 +25,7 @@ export * from './LoveButton'
 export * from './Title'
 export * from './SongBulkActions'
 export * from './useAlbumsPerPage'
+export * from './useGetHandleArtistClick.jsx'
 export * from './useInterval'
 export * from './useResourceRefresh'
 export * from './useToggleLove'
@@ -37,3 +38,4 @@ export * from './useRating'
 export * from './useSelectedFields'
 export * from './ToggleFieldsMenu'
 export * from './QualityInfo'
+export * from './formatRange.js'
